correção de security config

+++ src/main/java/br/com/boletos/repositories/BoletoRepository.java
@@ -9,7 +9,7 @@ import java.awt.print.Pageable;
 import java.util.List;
 import java.util.Optional;
 
-public interface BoletoRepository extends JpaRepository<Boleto, Long> {
+public interface BoletoRepository extends JpaRepository<Boleto, Integer> {
     List<Boleto> findByUuidAssociado(String uuid);
     List<Boleto> findByUuidAssociadoAndSituacaoBoleto(String uuid, String situacaoBoleto);
     Boleto findByIdBoletoAndDocumentoPagador(Integer idBoleto, String documentoPagador);
